Course_Design_of_Database_System_Principles/backend/requirements.txt
celery
django
django-debug-toolbar
djangorestframework
django-filter
django-cors-headers
django-extensions
mysqlclient
drf-spectacular
pillow 
pymysql